move extracted functions into utils module
and move useGetDepartures-related files into a separate folder
